Thêm Đăng nhập, Đăng xuất, Đổi mật khẩu

+++ CinemaBackend/src/main/java/com/example/CinemaBackend/entity/Payment.java
@@ -2,6 +2,8 @@ package com.example.CinemaBackend.entity;
 
 import jakarta.persistence.*;
 import lombok.Data;
+
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 @Data
@@ -17,7 +19,7 @@ public class Payment {
     private User user;
 
     @Column(nullable = false, precision = 10, scale = 2)
-    private Double amount;
+    private BigDecimal amount;
 
     @Enumerated(EnumType.STRING)
     @Column(name = "payment_method", nullable = false, length = 50)
@@ -73,14 +75,14 @@ public class Payment {
     /**
      * @return Double return the amount
      */
-    public Double getAmount() {
+    public BigDecimal getAmount() {
         return amount;
     }
 
     /**
      * @param amount the amount to set
      */
-    public void setAmount(Double amount) {
+    public void setAmount(BigDecimal amount) {
         this.amount = amount;
     }
